requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
django==5.0.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
inflection==0.5.1
    # via drf-yasg
packaging==23.2
    # via
    #   build
    #   drf-yasg
pip-tools==7.3.0
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via build
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
sqlparse==0.4.4
    # via django
uritemplate==4.1.1
    # via drf-yasg
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
